fix: logout button position
