+++ src/components/Profile.tsx
@@ -1,4 +1,5 @@
 import { useContext } from 'react'
+import { FiLogOut } from 'react-icons/fi';
 
 import { ChallengeContext } from '../contexts/ChallengeContext'
 import { useUsername } from '../contexts/UsernameContext';
@@ -15,6 +16,14 @@ export function Profile(props: ProfileProps) {
   const username = props.username
   const { level } = useContext(ChallengeContext)
 
+  const { removeUsername } = useUsername();
+  
+  function logout () {
+    window.confirm('Você realmente gostaria de deslogar?') &&
+    removeUsername()
+    window.location.reload();
+  }
+
   return(
       <div 
       className={styles.profileContainer}
@@ -22,6 +31,9 @@ export function Profile(props: ProfileProps) {
         <img src={`https://github.com/${username}.png`} alt={name}/>
         <div>
           <strong>{name}</strong>
+          <button onClick={logout}>
+            <FiLogOut/>
+          </button>
           <p>
             <img src='icons/level.svg' alt='level'/>
             Level {level}
