+++ src/pages/index.tsx
@@ -5,10 +5,8 @@ import { Profile } from '../components/Profile';
 import { ChallengeBox } from '../components/ChallengeBox';
 import { CountdownProvider } from '../contexts/CountdownContext';
 import { ChallengesProvider } from '../contexts/ChallengeContext';  
-import { useUsername } from '../contexts/UsernameContext';
 
 import Head from 'next/head';
-import { FiLogOut } from 'react-icons/fi';
 
 import styles from '../styles/pages/Home.module.css'
 
@@ -21,13 +19,7 @@ interface HomeProps {
 }
 
 export default function Home(props: HomeProps) {
-  const { removeUsername } = useUsername();
   
-  function logout () {
-    window.confirm('Você realmente gostaria de deslogar?') &&
-    removeUsername()
-    window.location.reload();
-  }
 
   return ( 
     <ChallengesProvider
@@ -41,9 +33,6 @@ export default function Home(props: HomeProps) {
       </Head>
 
       <ExperienceBar/>
-      <button onClick={logout}>
-        <FiLogOut/>
-      </button>
       <CountdownProvider>
       <section>
         <div>
